scripts/requirements.txt
google-api-python-client>=2.120
google-auth>=2.28
google-auth-oauthlib>=1.2
openai>=1.40
